[rpc] Mark get_transactions_in_range deprecated
## Description 

We will remove this endpoint soon, but it's currently used by the
Explorer. To prevent anyone else from using it, I am going to add
`deprecated` to the function name to discourage people from using the
method.

## Test Plan 

E2E

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes