Increase epoch length for all protocol upgrade tests
Tests were failing occasionally because one validator would not have its
vote sequenced before 2f+1 end of epoch. It is possible to fix this
reliably by monitoring the votes and then triggering end of epoch
manually, but that is quite a bit of work. Simply making the epochs
longer resolves the issue with very high probability.