Fix flaky committee change test
Previously we were checking for the execution of an arbitrtary
checkpoint to confirm that a node successfully made progress in joining
the committee. However, it is impossible to join the committee as a
validator without first executing all checkpoints in the previous epoch
and then reconfigging (as this changes the authority state to
validator). Therefore it is sufficient to check after some time that the
node is a validator and in the committee.

## Test Plan
Reproed against `4f5f730963` with 

```
MSIM_TEST_SEED=8146367765671037549 cargo simtest test_reconfig_with_committee_change_basic
```