[core] Ensure change_epoch tx is executed last
`change_epoch` tx is not causally ordered to be executed last, therefore
we must manually ensure this. In practice, this means that its effects
are persisted last.

Given the following:
1. checkpoints need not be executed in sequence number order, and
2. `TransactionManager` is responsible for persisting tx effects after
being scheduled by `CheckpointExecutor`

the most obviously correct solution is to do the following (which is the
approach taken in this PR):

1. At checkpoint execution time, if the checkpoint is the final of the
epoch, remove `change_epoch` tx from the list of txes enqueued to
`TransactionManager` (if not final checkpoint, nothing changes)
2. Once we have verified that all checkpoints but the last of the epoch
has been fully executed (and all other tx effects persisited), schedule
the `change_epoch` tx, and return to end the epoch. Note that
`check_epoch_last_checkpoint` checks if the
`highest_executed_checkpoint` watermark is the last of epoch, therefore
we choose this as our execution point.