[Narwhal] accept certificates in a separate task in Synchronizer
## Description 

Spawning off a task per certificate to avoid cancellations can be
inefficient and easy to cause leaks. Instead, have a dedicated task in
Synchronizer to process and accept certificates.

## Test Plan 

Deployed to private testnet.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes