Use blake2b everywhere
## Description 

Closing  

Previously, haneul address is computed as the first 32 bytes of the
SHA3-256 hash of `flag || pubkey`, this changes uses
first 32 bytes of the Blake2bhash of `flag || pubkey` instead. This is
for consideration of speed optimization.
## Test Plan 

All tests passes.

### Release notes
We now use the BLAKE2b-256 cryptographic hash function everywhere
instead of Sha3-256. It is considered to be just as secure as Sha3 but
is about three times faster. See also https://www.blake2.net.