Refactor haneul benchmark code for simplicity
## Description 

This PR does a few things to simplify adding new workloads:

1. Simplifications in workload_configuration.rs: There is no need to
handle individual workload in an error prone way to configure workload
weight, tps, etc. It only requires one call to the workload to configure
it now. For example, there is only one call to shared_counter workload
to configure it which looks like this:
```
let shared_workload = SharedCounterWorkloadBuilder::from(
            shared_counter_weight as f32 / total_weight as f32,
            target_qps,
            num_workers,
            in_flight_ratio,
            shared_counter_hotness_factor,
        );
```
2. Simplification in generating gas: There is no need to handle
individual workload while generating gas. The call to generate gas is
handled by a trait now in:
```
pub async fn generate(
        &mut self,
        builders: Vec<Box<dyn WorkloadBuilder<dyn Payload>>>,
        gas_price: u64,
        chunk_size: u64,
    ) -> Result<Vec<Box<dyn Workload<dyn Payload>>>> {
```
3. Remove the concept of combination workload as it was hard to maintain
while it was not being used in production anymore. This helped remove a
lot of code which simplified the crate a lot.
4. Structs like `WorkloadInitGas`, `WorkloadPayloadGas`, etc are all
removed now
5. `Payload` and `Workload` abstractions are much cleaner
## Test Plan 

Existing tests