[Rocksdb] remove unused arguments and move overrides into typed_store
## Description 

Remove `default_db_options()` arguments that are basically unused. The
only overrides to these options are for index storage, which I believe
is going away. In any case, it seems better to override at callsites
instead of passing in overrides as arguments.

Move the overrides in `default_db_options()` into the base
`default_db_options()` in typed store. This allows sharing the settings
with Narwhal.

## Test Plan 

n/a

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes