Haneul orchestrator
This PR provides facilities to quickly deploy and benchmark the codebase
in a geo-distributed environment. It serves a very different purpose
than the PE-maintained clusters. Here the goal is to quickly produce
L-graphs under various systems configuration (various number of nodes,
crash-faults patterns, etc) on custom cluster. Each user can deploy its
(temporary) cluster and benchmark various changes of the codebase as
quickly as possible.

The PR main look large but most of the code is not critical (eg, it
boots nodes and execute ssh commands). I would however suggest to
carefully review the following files:
*
https://github.com/GeunhwaJeong/haneul/blob/orchestrator/crates/haneul-orchestrator/src/measurement.rs
*
https://github.com/GeunhwaJeong/haneul/blob/orchestrator/crates/haneul-orchestrator/assets/plot.py

We made an effort to document the new crate as clearly as possible and
the following readme provides a step-by-step guide to use it:
https://github.com/GeunhwaJeong/haneul/blob/orchestrator/crates/haneul-orchestrator/readme.md.
Feel free to ping me if you need access to our research AWS account.