[fix] connectivity monitor metrics
## Description 

Connectivity monitor metrics seemed to be broken, as the `network_peers`
metrics was always getting increasing. The PR is fixing the issue
ensuring that the `decr` gets called whenever a disconnection happens +
introduces some testing to ensure we won't accidentally break things as
metrics appear to be quite important.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes