Few fixes for gas charging
## Description 

Few gas fixes. We are not charging for gas computation any longer as the
buckets, the initial read charge and the native charging for dynamic
loading take care of it.
Also making the computation of buckets correct for the last bucket. On
the last bucket it may happen that what was charged up to that point was
more than the max charge in the bucket and so that would result in a
overflow. We now check for it and if we underflow we do nothing which
translate to charging as much as we have up to that point

## Test Plan 

Existing tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes