[tests][programmable transactions] Add tests for MakeMoveVec
## Description 

- Added transactional tests for objects
- Added "generated" tests for primitive types
- BCS bytes are now always checked for `MakeMoveVec` primitive
arguments. And checked before args reach the VM. This isn't strictly
necessary for safety, but provides a more predicable UX.

## Test Plan 

- It is tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [X] breaking change for FNs (FN binary must upgrade)
- [X] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [X] necessitate either a data wipe or data migration

### Release notes