[framework] Kiosk+
## Description 

This is a new iteration of  and .

### Module haneul::transfer_policy

A module for creating authorized transfers with a composable and very
extendable interface. Used by Kiosk by default on all purchase calls.
Allows for any combination of custom "Rules" which must be completed for
the TransferRequest to be confirmed.

This is the first example of the "Crisps" pattern - a hot potato with a
Bag inside - or a bag of hot potato. 😛

### Module haneul::kiosk

Very simple trading primitive which allows for two main actions:

- `kiosk::list` - lists an item for X amount of HANEUL
- `kiosk::list_with_purchase_cap` - create a PurchaseCap for an item
with some min_price in HANEUL

Kiosk enforces TransferPolicy on all trades; which also means that a
policy needs to be issued for a type to become tradable in Kiosks.

## Test Plan 

All modules tested.
All of the features covered.
Edge cases are yet to be discovered (but many are already tested).

### Type of Change (Check all that apply)

- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

- Adds `haneul::kiosk` module which allows setting up a "Kiosk" for p2p
trades.
- Adds `haneul::transfer_policy` module which allows creators to set custom
trading policies for their items (such as royalty enforcement).
-

---------