[Narwhal] reduce timestamp drift tolerance to 1s
## Description 

Narwhal headers are produced at 0.5s ~ 1s intervals. Using a larger
clock drift tolerance would just suspend the header until it gets
cancelled, making it harder to debug. So using a smaller drift which
should be more obvious to discover.

In future, we can implement a Narwhal clock that removes the need for
operators to detect and fix the issue on their own.

## Test Plan 

Verify in a geo distributed network.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------