[ts-sdk] Normalize Haneul Address in more places
## Description 

This normalizes Haneul address in more cases in the TS SDK, so that they
don't require normalization as much externally. Specifically
`toHaneulAddress()` in public keys, and when taking an object as input in
the programmable transaction builder. The effect of this is that `0x`
should now consistently be included in address internally.

## Test Plan 

Updated tests to include `0x`.