[wallet-ext] open Ledger connection modal in new tab to prevent HID/USB connection issues
## Description 
TLDR; Some browser APIs don't work properly when an extension is in
popup mode, so we have to re-work some of the Ledger flow UX. The first
UX change is that we need to always open the Ledger connection/import
flow in a new window, which is what this PR changes. The second UX
change (will tackle this in a follow-up) is that we need to stop
requesting user connections during the signing flow which is something
that I introduced yesterday.

https://capture.dropbox.com/ArbF7mmK4Ws1OkYj

### More information re: the root bug
`ledger/transport-web-hid` relies on the `HID.requestDevices` API
implemented in Chrome, and you can only call this API in the context of
a user gesture (mousedown, keydown, etc.). The API uses this logic
(https://developer.mozilla.org/en-US/docs/Web/API/UserActivation) to
determine whether or not the API call fits in the context of a user
gesture, but the logic simply doesn't work as expected when the Chrome
extension is in popup mode.

## Test Plan 
- Manual testing
  - Can connect to Ledger successfully 100% of the time now
- Despite demo'ing this several times, I can't seem to connect to Ledger
in popup mode whatsoever now

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes