[Narwhal] call handlers directly for local PrimaryToWorker and WorkerToPrimary communications
## Description 

We have observed the following issues in practice:
1. PrimaryToWorker and WorkerToPrimary connections sometimes break or
stay broken. The can be multiple causes, e.g. network misconfigurations,
load, fragile logic or something else.
2. There are components that are dependencies for network handlers, or
part of consensus, e.g. Synchronizer and Subscriber, that need access to
the network, but passing Network to them cannot be done at creation
time.

This change aims to address the above two issues:
1. Wire PrimaryToWorker and WorkerToPrimary handlers to the client
callsites directly, without going through the networking layer.
2. Pass a `NetworkClient` object to components that need access to the
network. Local handlers and in future remote Networks will be wired to
the `NetworkClient`, but these do not need to happen before the creation
of `NetworkClient`.



## Test Plan 

existing tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes