[dry run][dev inspect] Fix dynamic field edge case
## Description 

Set gas objects in dry run and dev-inspect to the max version to avoid
dynamic field issues we have seen with faucet (and faucet tests).
In dry run and dev inspect, you might load a dynamic field that is
actually too new for the transaction. Ideally, we would want to load the
"correct" dynamic fields, but as that is not easily determined, we
instead can set the sequence number of the gas objects to MAX - 1, to
ensure that the lamport version for the transaction is MAX, and as such,
a valid lamport version for any object used in the transaction
(preventing internal assertions or invariant violations from being
triggered)

## Test Plan 

- Added new tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes