[explorer] instrument PostHog analytics to give it a try
## Description 

This PR instruments PostHog analytics on the Explorer since we're
testing out some new analytic providers. If we exceed the free volume,
then PostHog will lock our access but still capture all of our events so
we can get an estimate on volume. I've also gated this behind a feature
flag in case something goes wrong or we want to turn it off.

## Test Plan 
- Events are flowing into PostHog
- Tested with the feature flag ON/OFF

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A