[Narwhal] initialize primary->worker handler with Network
## Description 

Worker's PrimaryReceiverHandler was a dependency for creating Network,
so it could not have a Network member variable in a straightforward way.
But Network is needed to send outgoing RPCs. The workaround was to get
the Network object from the incoming request.

Now for handling local connections, Network object is already available
when constructing the primary->worker local handler, but no longer
available as part of the request. So we store Network in
PrimaryReceiverHandler optionally, when it is used as a local handler.

## Test Plan 

unit test

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes