[narwhal] ensure sub dag timestamp is monotonically incremented
## Description 

This PR is enforcing the sub dag's commit timestamp to always
monotonically increase. Basically we enforce the update
`max(previous_sub_dag_commit_timestamp, current_leader_timestamp`.

The motivation behind this change is to ensure extra-safety on our
timestamps and avoid any potential downstream issues.

Took the opportunity on this PR to refactor the code a little bit and
simplify the information we keep in the consensus state to mainly use
the last committed sub dag.

## Test Plan 

Introduced unit tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes