[checkpoints] Backward compatibility for builder_checkpoint_summary table
Since we made a decision to keep live updates for testnet, this PR adds
backward compatibility for the `builder_checkpoint_summary` table.

The new table, `builder_checkpoint_summary_v2` is added, and the logic
is adjusted to account for both tables present.

When a validator switches to a new version, it stops deleting pending
checkpoints, and instead switches to reading indexes in a v2 table.

Because of this behavior we luckily do not need a feature flag for this
change. Reverting validator code however won't be possible after this
upgrade.