[apps] Tweak lint rules to disable shorthand conversion
## Description

This is a follow-up to , making it required for us to use
`Number()` to convert things to numbers so that we can easily find these
in our code. I also removed shorthand string conversion because that's
ugly anyway. I kept shorthand booleans because they're less error-prone
and easier to find (just look for `!!`).

#### Why?

We recently moved a lot more of our API to return strings, and depend on
a lot more string -> number conversion in our apps. The problem is that
the string encoded values are u64 (or higher), which isn't representable
as numbers. We don't expect any immediate issues to arise from this but
long term they will pop up, and by making this conversion easy and
obvious to find (just search for `Number(`), we make it easy to fix this
across our codebase.

## Test Plan

Non-functional change things should still work.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes