[Compatibility] A few small tweaks
## Description 

* Fix `sed` commands
* Disable incremental build (which will be slow for our purposes)
* Add debug log level selector for `workflow_dispatch` 

## Test Plan 

Must merge before it can be tested, otherwise we don't have access to
the script changes on the vultr host.

Manually tested the new sed commands locally on the vultr host in ssh
session. Should be good.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes