Add bounds for iterators in rocksdb
## Description 

This PR allows the caller to pass iterator lower and upper bounds
upfront (when creating the iterator) which can be useful for db in
preventing unneeded key scans. Also, start using bounds for
`get_dynamic_fields_iterator()` which was the cause for recent incident.
We will add this to more callsites with composite keys i.e. (ObjectID,
..) for example
 
## Test Plan 

Added tests