[Compatibility] Fix clean mode default value
## Description 

[Most recent
run](https://github.com/GeunhwaJeong/haneul/actions/runs/4687467829/jobs/8306807903)
of the workflow failed due to disk being full from previous runs. The
job is supposed to wipe the haneul db as a step unless explicitly
specified. But this step was skipped.

I think this is because I introduced that conditional argument (`clean`)
only under `workflow_dispatch`, so for a scheduled run, the boolean is
undefined, and probably evaluates to false in that case.

The fix here is to invert the logic of the boolean so that if true it
will not be wiped, and it will be true only when explicitly set (in
dispatch mode).

## Test Plan 

Added `pull_request:` dispatch mode temporarily, without input
arguments, pushed commit, and verified that db wipe step was not
skipped:

![Screenshot 2023-04-13 at 7 45 51
AM](https://user-images.githubusercontent.com/11576329/231798029-147249ac-85a5-4793-b51c-d72f095b3b52.png)

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes