[Compatibility] Run against latest binary build that passed CI
## Description 

This allows for 3 things:
* Faster run by avoiding long release builds
* Ensuring that we are checking against a known good commit from the
perspective of CI (which makes bisecting and investigating issues easier
as there are not any commits in the check that have not also gone
through the rest of CI)
* Consistency of run across many machines when we run a sharded
configuration

## Test Plan 

Temporarily enabled `pull_requests` dispatch to get a CI run. Validated
that all steps succeeded (or at least successfully kicked off in the
case of the final one which takes many hours):
https://github.com/GeunhwaJeong/haneul/actions/runs/4695597529/jobs/8324912485

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes