[Compatibility] Use snapshots to shard
## Description 

This PR uses a matrix to run across n hosts, with all but one hosts
restoring from snapshot. This should help with runtime and disk
utilization.

## Test Plan 

Temporarily set `pull_request` trigger to test in CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes