[CompatibilityChecker] Cleanup orphan processes for cancelled jobs
## Description 

When a job is cancelled, e.g. due to another job in the workflow
failing, the job is marked as cancelled and the sync script and haneul-node
process are orphaned. These remain bound to the haneul-node ports, and thus
cause the next run attempt to fail.

This adds a step to kill the orphaned processes. Note that in the fail
case, either the monitor script notices that the haneul-node is no longer
responding to polls, or the monitor script itself fails. In either case,
cleanup is run by trapping EXIT.

## Test Plan 

👀 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes