[refactor] randomized tests to run in multithreaded environment + speed up
## Description 

Converting the randomized tests suite to run in a multithreaded
environment so we can speed up test execution. We currently [see in our
CI](https://github.com/GeunhwaJeong/haneul/actions/runs/4728135879/jobs/8389393727#step:6:2903)
the tests timing out daily as they are taking lots of time.

However, I've identified that clearing the database between every test
iteration was a quite costly operation (locally measuring ~8 seconds)
which is extremely prohibitive and primary source of great delay. In the
end of the day is unnecessary as we only read from the store when we
recover so it doesn't harm not clearing the database across execution.
Haven't looked deeper why dropping a column family in RocksDB takes that
much, but for now bypassing works equally fine.

We now create the database only once and re-use across all the tests
without clearing. Only because of this change `I am looking to a speed
up of ~15x`. Any additional benefits from the multithreaded environment
will be extra here.

Once I confirm on next run that we indeed have significant gains I'll
adjust the properties accordingly so we test in much more iterations.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes