[wallet-ext] fix tooltips being hidden behind overlay screens
## Description 
Tooltips in the wallet are portaled to the body, and the tooltips
started getting hidden after the recent changes to the layout of the
Overlay component. This PR just increments the z-index to `99999` to
match that of the `Toaster` (the Overlay uses `9999`). We probably need
to do some better z-index management in the future to prevent issues
like this one from happening, but that will be for another day :)

<img width="362" alt="image"
src="https://user-images.githubusercontent.com/7453188/233657695-d03a26b0-dfd2-4fcc-9ca3-51566d10cc18.png">


## Test Plan 
- Manual testing

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A