[tx-fuzzer] use one AuthorityState throughout test cases
## Description 

This PR adds a `run_proptest` function that runs a provided test case,
with given number of times, and input generated with the provided
strategy. This attempts to do what `proptest` macro does but with the
same initialized `AuthorityState` throughout. `gas_data_tests` has been
changed to use this function.
  
## Test Plan 

It's all tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes