Optimize transaction construction, especially for stake transactions
## Description 

This does two things primarily:
- Reorganize some of the logic around data fetching when building
transactions so that we fetch slightly more concurrently. There's still
more optimization opportunity here but it's diminishing returns for
increased effort.
- Optimize stake requests in the wallet, which removes 2 serialized API
calls that were required to serialize the transaction to BCS. In testing
this removes up to 1s of latency in the stake flow.

## Test Plan 

Ran locally.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes