[rpc] Fix get_coin_metadata for 0x2::haneul::HANEUL
## Description 

we assumed
[previous_transaction](https://github.com/GeunhwaJeong/haneul/blob/main/crates/haneul-json-rpc/src/coin_api.rs#L99-L102)
is always the publish transaction for a package, but this assumption has
been broken by the package upgrades. For system objects, the package ids
can stay unchanged even though their content can be changed by
`ChangeEpoch` transaction.

In this case, we can fetch the genesis transaction directly, which is
the transaction that creates the coin metadata object for HANEUL

This PR also adds memoization for the `get_package_object_id` function

## Test Plan 

```
curl --location --request POST 'http://127.0.0.1:9000' \
--header 'Content-Type: application/json' \
--data-raw '{
    "jsonrpc":"2.0",
    "id":1,
    "method":"haneulx_getCoinMetadata",
    "params":["0x2::haneul::HANEUL"]
}'
```

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes