Use coin symbol from metadata
## Description 

Our apps were using a parsed coin symbol from the type name, not from
the metadata. This updates to only use the parsed symbol as a fallback,
and use the coin metadata when we can. The one side-effect of this is
that symbol have a small delay when loading, but they will at least
display the correct values now.

## Test Plan 

Ran explorer locally

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes