[refactor] validator to submit with configurable position and delay step
## Description 

Providing the ability to configure the maximum position
(`max_submit_position`) of a validator to submit to consensus. Provided
a max position `N`, we guarantee that a validator even if is elected to
submit to position `M` , where `M > N`, it will submit on position `N`.

Similar, if the `submit_delay_step_millis` property is set, then we
ensure that on the consensus adapter the delay step will get overwritten
with this constant value instead of using the latency observer logic.

## Test Plan 

Added unit tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes