Fix loading incomplete NFTs
## Description 

Describe the changes or additions included in this PR.


- Fix wallet loading partial NFT by fetching and filtering for NFTs
(10). Scroll to load more
- Move `useGetOwnedObjects` and `useOnScreen` to core
- Added map key to BalanceChangeEntry

## Test Plan 


https://user-images.githubusercontent.com/126525197/235388841-bae92114-d54b-4efd-87f0-7afbedb64478.mov


How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes