[wallet-ext] add Ledger toast notification that's displayed when you open the wallet
## Description 
We're adding a toast to the wallet token page notifying users of the new
Ledger integration feature. The idea is that we're going to show this
toast to every user until they manually dismiss it by clicking the "X"
button or tapping the notification itself. I want to note that this is a
temporary feature, and we definitely shouldn't be using toasts for
notifications like this in the future. This was the quickest approach vs
rendering a fake toast-looking notification overlay which has its own
problems (e.g., how do you handle the stacking animation if you have a
toast and the toast-looking notification?)

https://capture.dropbox.com/pfo9to23ZWWUBt5f?autoPreview=1

## Test Plan 
- Manual testing (switching tabs, having another toast pop-up on the
token page, opening the account menu, etc.)
- Tested Plausible logging works
- Tested with feature flag ON and OFF

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A