Fix wait for local execution bug in sdk
This PR fixes a bug in the SDK where it doesn't properly wait for tx
execution on fullnodes. Specifically, it returns when it sees the
transaction, but that doesn't guarantee that the transaction was
actually executed.
In the fix we simply use retry instead of timeout because fullnode local
execution already uses a timeout to wait.
This fixes a flaky test in haneul-surfer.